[REF] Move class `PETPipeline` to new `clinica.pipelines.pet` module (#1223)
